requirements.txt
asgiref==3.2.7
Django==3.0.5
django-crispy-forms==1.9.0
numpy==1.18.2
Pillow==7.1.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
